src-tauri/scripts/requirements.txt
python-socketio[client]>=5.11.0
websocket-client>=1.9.0
pyinstaller>=6.15.0
